explicitly set default results to $PWD/results
this change affects people running the workflow directly from
github, e.g.

$ nextflow run pgscatalog/pgsc_calc ...

if --outdir isn't set, then the results folder can be in $NXF_HOME,
which is a hidden folder in the home directory by default. not a
helpful place for results to be!

this doesn't affect people running from a cloned repo directly